requirements.txt
blinker==1.9.0
click==8.1.8
colorama==0.4.6
Flask==3.1.0
Flask-JWT-Extended==4.7.1
Flask-SQLAlchemy==3.1.1
mysql-connector-python==9.1.0
python-dotenv==1.0.1
Werkzeug==3.1.3
SQLAlchemy==2.0.36